backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.0.3
langchain==0.0.340
langchain-community==0.0.11
langchain-openai==0.0.3
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
networkx==3.2.1
python-multipart==0.0.6
pypdf==3.17.1
docx2txt==0.8
python-docx==0.8.11
reportlab==4.0.7
matplotlib==3.8.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
faiss-cpu==1.7.4
dashscope==1.14.1